refactor(encoding): relocate internal/infrastructure/encoding to internal/pkg/encoding (#1336)
Move the locale-decoding utility to the shared internal/pkg home (ADR-062 Decision 1). Zero content change: build variants and white-box test intact. Update the three importers (pipeline.go, process_executor.go, executor.go).
