+++ internal/pkg/encoding/encoding.go
