+++ internal/pkg/encoding/encoding_other.go
