+++ internal/pkg/encoding/encoding_windows.go
